Permit multiple pci-root controllers for mac99
We should not use pci.0 (the AGP bus) or pci.1 (internal bus),
so we're left with pci.2 (the main bus that things should actually utilize).

However, this is currently hardcoded to only be allowed for pSeries
machines. We can add ourselves in temporarily while this is being
handled upstream.

See also:
 - https://gitlab.com/qemu-project/qemu/-/issues/1146
 - https://gitlab.com/libvirt/libvirt/-/issues/360